tidy: Fix "multiplication implicit widening" in mytharchive.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

1) Cast multiplication arguments to a larger type.

2) Convert a variable to the appropriate type.

3) Remove a cast from int64_t to int, that then gets implicitly cast back
to int64_t.

+++ mythplugins/mytharchive/mytharchive/thumbfinder.cpp
@@ -388,7 +388,7 @@ void ThumbFinder::updateThumb(void)
     item->SetImage(imageFile, "", true);
 
     // update the image grid item
-    int64_t pos = (int) ((m_currentPTS - m_startPTS) / m_frameTime);
+    int64_t pos = (m_currentPTS - m_startPTS) / m_frameTime;
     thumb->frame = pos - m_offset;
     if (itemNo != 0)
     {
